% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_sizing.R
\name{calculate_ballon_volume_auto}
\alias{calculate_ballon_volume_auto}
\title{Auto-size thermal storage volume}
\usage{
calculate_ballon_volume_auto(p_pac_kw, cop_nominal, t_tolerance)
}
\arguments{
\item{p_pac_kw}{Heat pump electrical power (kW)}

\item{cop_nominal}{Nominal COP at reference temperature}

\item{t_tolerance}{Temperature tolerance (degrees C). The usable
temperature range is \code{2 * t_tolerance}.}
}
\value{
Tank volume in litres (rounded to nearest 50 L)
}
\description{
Calculates the recommended tank volume to store 2 hours of heat pump
output within the temperature tolerance band.
}
\details{
Formula: \code{V = P_PAC * COP * 2h / (2 * tolerance * 0.001163)},
rounded to the nearest 50 L.
}
\examples{
calculate_ballon_volume_auto(60, cop_nominal = 3.5, t_tolerance = 5)
# => 36400 L (simplified example)
}
