% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_sizing.R
\name{calculate_pv_auto}
\alias{calculate_pv_auto}
\title{Auto-size PV installation}
\usage{
calculate_pv_auto(p_pac_kw, volume_ballon_l, cop_nominal, t_consigne = 55)
}
\arguments{
\item{p_pac_kw}{Heat pump electrical power (kW)}

\item{volume_ballon_l}{Tank volume in litres (used for ECS estimation)}

\item{cop_nominal}{Nominal COP}

\item{t_consigne}{Target tank temperature (degrees C, default 55)}
}
\value{
PV capacity in kWc (rounded to nearest 0.5 kWc)
}
\description{
Estimates the optimal PV capacity to cover annual heat pump electricity
consumption, based on an average Belgian yield of 950 kWh/kWc/year.
}
\details{
Two sizing strategies depending on heat pump power:
\describe{
  \item{P_PAC <= 10 kW (domestic ECS)}{Sizes based on daily hot water
    demand + standing losses.}
  \item{P_PAC > 10 kW (collective/heating)}{Sizes based on 5 equivalent
    full-load hours per day.}
}
}
\examples{
calculate_pv_auto(60, volume_ballon_l = 2000, cop_nominal = 3.5)
}
