% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_co2_elia.R
\name{compute_co2_impact}
\alias{compute_co2_impact}
\title{Calcule l'impact CO2 de l'optimisation}
\usage{
compute_co2_impact(sim, co2_15min)
}
\arguments{
\item{sim}{DataFrame de simulation avec offtake_kwh, sim_offtake, timestamp}

\item{co2_15min}{Vecteur d'intensite CO2 au pas 15 min (gCO2eq/kWh)}
}
\value{
list avec les KPIs et vecteurs horaires
}
\description{
Calcule l'impact CO2 de l'optimisation
}
