% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_openmeteo.R
\name{interpolate_temperature_15min}
\alias{interpolate_temperature_15min}
\title{Interpole les temperatures horaires au pas de 15 min}
\usage{
interpolate_temperature_15min(df_hourly, ts_15min)
}
\arguments{
\item{df_hourly}{tibble avec colonnes timestamp et t_ext (horaire)}

\item{ts_15min}{Vecteur de timestamps au pas de 15 min}
}
\value{
Vecteur de temperatures interpolees (meme longueur que ts_15min)
}
\description{
Interpole les temperatures horaires au pas de 15 min
}
