% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_optimizer.R
\name{MILPOptimizer}
\alias{MILPOptimizer}
\title{MILP Optimizer}
\description{
R6 optimizer using Mixed Integer Linear Programming.
  Delegates block solving to solve_block() from R/optimizer_milp.R.
}
\section{Super class}{
\code{wisepocpac::BaseOptimizer} -> \code{MILPOptimizer}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MILPOptimizer-clone}{\code{MILPOptimizer$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="wisepocpac" data-topic="BaseOptimizer" data-id="get_results"><a href='../../wisepocpac/html/BaseOptimizer.html#method-BaseOptimizer-get_results'><code>wisepocpac::BaseOptimizer$get_results()</code></a></span></li>
<li><span class="pkg-link" data-pkg="wisepocpac" data-topic="BaseOptimizer" data-id="guard_baseline"><a href='../../wisepocpac/html/BaseOptimizer.html#method-BaseOptimizer-guard_baseline'><code>wisepocpac::BaseOptimizer$guard_baseline()</code></a></span></li>
<li><span class="pkg-link" data-pkg="wisepocpac" data-topic="BaseOptimizer" data-id="initialize"><a href='../../wisepocpac/html/BaseOptimizer.html#method-BaseOptimizer-initialize'><code>wisepocpac::BaseOptimizer$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="wisepocpac" data-topic="BaseOptimizer" data-id="solve"><a href='../../wisepocpac/html/BaseOptimizer.html#method-BaseOptimizer-solve'><code>wisepocpac::BaseOptimizer$solve()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MILPOptimizer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MILPOptimizer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MILPOptimizer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
