% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_optimizer.R
\name{SmartOptimizer}
\alias{SmartOptimizer}
\title{Smart Optimizer (Rule-based)}
\description{
R6 optimizer using the rule-based Smart mode from app.R.
  Contains the decider() and run_simulation() logic. Unlike the mathematical
  optimizers, this processes one quarter-hour at a time sequentially.
}
\section{Super class}{
\code{\link[wisepocpac:BaseOptimizer]{wisepocpac::BaseOptimizer}} -> \code{SmartOptimizer}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SmartOptimizer-solve}{\code{SmartOptimizer$solve()}}
\item \href{#method-SmartOptimizer-clone}{\code{SmartOptimizer$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="wisepocpac" data-topic="BaseOptimizer" data-id="get_results"><a href='../../wisepocpac/html/BaseOptimizer.html#method-BaseOptimizer-get_results'><code>wisepocpac::BaseOptimizer$get_results()</code></a></span></li>
<li><span class="pkg-link" data-pkg="wisepocpac" data-topic="BaseOptimizer" data-id="guard_baseline"><a href='../../wisepocpac/html/BaseOptimizer.html#method-BaseOptimizer-guard_baseline'><code>wisepocpac::BaseOptimizer$guard_baseline()</code></a></span></li>
<li><span class="pkg-link" data-pkg="wisepocpac" data-topic="BaseOptimizer" data-id="initialize"><a href='../../wisepocpac/html/BaseOptimizer.html#method-BaseOptimizer-initialize'><code>wisepocpac::BaseOptimizer$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SmartOptimizer-solve"></a>}}
\if{latex}{\out{\hypertarget{method-SmartOptimizer-solve}{}}}
\subsection{Method \code{solve()}}{
Run the smart (rule-based) simulation.
  Overrides BaseOptimizer$solve() completely since the Smart mode
  doesn't use the block loop pattern.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SmartOptimizer$solve()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The dataframe with simulation columns added
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SmartOptimizer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SmartOptimizer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SmartOptimizer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
