% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_co2_elia.R
\name{fetch_co2_intensity}
\alias{fetch_co2_intensity}
\title{Recupere l'intensite CO2 horaire pour une plage de dates}
\usage{
fetch_co2_intensity(start_date, end_date)
}
\arguments{
\item{start_date}{Date de debut}

\item{end_date}{Date de fin}
}
\value{
list(df = tibble(datetime, co2_g_per_kwh), source = character)
}
\description{
Recupere l'intensite CO2 horaire pour une plage de dates
}
