% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co2_elia.R
\name{interpolate_co2_15min}
\alias{interpolate_co2_15min}
\title{Interpole l'intensite CO2 horaire au pas de 15 min}
\usage{
interpolate_co2_15min(co2_hourly, ts_15min)
}
\arguments{
\item{co2_hourly}{tibble avec colonnes datetime et co2_g_per_kwh (horaire, UTC)}

\item{ts_15min}{Vecteur de timestamps au pas de 15 min (Europe/Brussels)}
}
\value{
Vecteur d'intensite CO2 interpolee (meme longueur que ts_15min)
}
\description{
Interpole l'intensite CO2 horaire au pas de 15 min
}
